Move content compilation into separate standalone script (#97)
In order to accommodate building the Advisor content in a separate
container that should not need to install the entire Advisor code base,
we separate out the 'compile' and optional 'dump' code into a self
contained script, `compile_advisor_content.py`.  The Django management
command `import_content` code then imports those functions, to avoid
code duplication but to keep its ability to process either the dumped
content files or the raw repositories (which are what our tests are
based on).

The playbook repository path defaults to the content repository path
if the `-p` option is not specified.

SourceryAI spotted that we weren't matching the paths correctly in
finding the git hash for a playbook (which we use for the `version` field).
ClaudeAI helped me fix this!

SourceryAi also suggested making sure playbook file is closed, actually
using the filename constants we declared, and some further code
simplifications.
