+++ api/advisor/api/tests/test_import_content_command.py
@@ -304,6 +304,8 @@ class ImportContentTestCase(TestCase):
             active_rule.playbooks()[0].description,
             'Fix for Active_rule on rhel/host'
         )
+        # Playbook version hash should be matched
+        self.assertIsNotNone(active_rule.playbooks()[0].version)
 
         # Then test the differences
         acked_rule = Rule.objects.get(rule_id=constants.acked_rule)
@@ -905,7 +907,7 @@ class CoverageTestCase(TestCase):
 
     def test_read_playbook_failures(self):
         playbook_file = join(PATH_TO_TEST_CONTENT_REPO, 'playbook.yaml')
-        from api.management.commands.import_content import read_playbook
+        from api.scripts.compile_advisor_content import read_playbook
         with FileDeleter(playbook_file):
             # First: no name but first task is named
             playbook_content = [{
@@ -960,7 +962,8 @@ class CoverageTestCase(TestCase):
             playbook_3 = playbooks[rule_fix]
             self.assertEqual(playbook_3['description'], 'step 3')
             self.assertEqual(playbook_3['play'], '- name: step 3\n')
-            self.assertEqual(playbook_3['path'], '/playbooks/rhel_host/fixit.yml')
+            # No git repository = full path including temp_dir.
+            self.assertEqual(playbook_3['path'], f'{temp_dir}/playbooks/rhel_host/fixit.yml')
             # rule_id contains the temp_dir name
             self.assertEqual(playbook_3['rule_id'], rule_id)
             self.assertEqual(playbook_3['type'], 'fix')
